+++ src/index.ts
@@ -1,6 +1,8 @@
 import { BufferReader, BufferWriter } from "./buffer";
-import { DeckDefinition, DeckCard } from "../types";
+import { DeckDefinition, DeckCard, SideboardCard } from "../types";
 import { DECKSTRING_VERSION, FormatType } from "./constants";
+
+type BaseCard = [number, number, ...any[]];
 
 function verifyDbfId(id: unknown, name?: string): void {
 	name = name ? name : "dbf id";
@@ -19,17 +21,16 @@ function isPositiveNaturalNumber(n: unknown): boolean {
 	return n > 0;
 }
 
-function sort_cards<T extends DeckCard>(cards: T[]): T[] {
+function sort_cards<T extends BaseCard>(cards: T[]): T[] {
 	return cards.sort((a, b) => (a[0] < b[0] ? -1 : a[0] > b[0] ? 1 : 0));
 }
 
-function trisort_cards<T extends DeckCard>(cards: T[]): [T[], T[], T[]] {
+function trisort_cards<T extends BaseCard>(cards: T[]): [T[], T[], T[]] {
 	const single: T[] = [],
 		double: T[] = [],
 		n: T[] = [];
 	for (const tuple of cards) {
 		const [card, count] = tuple;
-		verifyDbfId(card, "card");
 		if (count === 0) {
 			continue;
 		}
@@ -55,7 +56,9 @@ export function encode(deck: DeckDefinition): string {
 			deck.format !== FormatType.FT_STANDARD &&
 			deck.format !== FormatType.FT_CLASSIC) ||
 		!Array.isArray(deck.heroes) ||
-		!Array.isArray(deck.cards)
+		!Array.isArray(deck.cards) ||
+		(typeof deck.sideboardCards !== "undefined" &&
+			!Array.isArray(deck.sideboardCards))
 	) {
 		throw new Error("Invalid deck definition");
 	}
@@ -65,6 +68,7 @@ export function encode(deck: DeckDefinition): string {
 	const format = deck.format;
 	const heroes = deck.heroes.slice().sort();
 	const cards = sort_cards(deck.cards.slice());
+	const sideboard = sort_cards((deck.sideboardCards || []).slice());
 
 	writer.null();
 	writer.varint(DECKSTRING_VERSION);
@@ -79,11 +83,31 @@ export function encode(deck: DeckDefinition): string {
 		writer.varint(list.length);
 		for (let tuple of list) {
 			const [card, count] = tuple;
+			verifyDbfId(card), "card";
 			writer.varint(card);
 			if (count !== 1 && count !== 2) {
 				writer.varint(count);
 			}
 		}
+	}
+
+	if (sideboard.length) {
+		writer.varint(1);
+		for (let list of trisort_cards(sideboard)) {
+			writer.varint(list.length);
+			for (let tuple of list) {
+				const [card, count, owner] = tuple;
+				verifyDbfId(card, "sideboard card");
+				verifyDbfId(owner, "sideboard card owner");
+				writer.varint(card);
+				if (count !== 1 && count !== 2) {
+					writer.varint(count);
+				}
+				writer.varint(owner);
+			}
+		}
+	} else {
+		writer.varint(0);
 	}
 
 	return writer.toString();
@@ -96,7 +120,7 @@ export function decode(deckstring: string): DeckDefinition {
 		throw new Error("Invalid deckstring");
 	}
 
-	const version = reader.nextVarint();
+	const version = reader.nextByte();
 	if (version !== DECKSTRING_VERSION) {
 		throw new Error(`Unsupported deckstring version ${version}`);
 	}
@@ -120,15 +144,31 @@ export function decode(deckstring: string): DeckDefinition {
 	for (let i = 1; i <= 3; i++) {
 		for (let j = 0, c = reader.nextVarint(); j < c; j++) {
 			cards.push([
-				reader.nextVarint(),
+				reader.nextVarint(), // dbf id
 				i === 1 || i === 2 ? i : reader.nextVarint(),
 			]);
 		}
 	}
 	sort_cards(cards);
 
+	const sideboardCards: SideboardCard[] = [];
+	const hasSideboard = reader.nextByte();
+	if (hasSideboard == 1) {
+		for (let i = 1; i <= 3; i++) {
+			for (let j = 0, c = reader.nextVarint(); j < c; j++) {
+				sideboardCards.push([
+					reader.nextVarint(), // dbf id
+					i === 1 || i === 2 ? i : reader.nextVarint(),
+					reader.nextVarint(), // sideboard card owner dbf id
+				]);
+			}
+		}
+		sort_cards(sideboardCards);
+	}
+
 	return {
 		cards,
+		sideboardCards,
 		heroes,
 		format,
 	};