+++ types.d.ts
@@ -1,11 +1,13 @@
 export type FormatType = 1 | 2 | 3;
 
 export type DeckCard = [number, number]; // [dbfId, count]
-
 export type DeckList = DeckCard[]; // keep type for backwards compatibility
+
+export type SideboardCard = [number, number, number]; // [dbfId, count, sideboardOwnerDbfId]
 
 export interface DeckDefinition {
 	cards: DeckCard[];
+	sideboardCards: SideboardCard[];
 	heroes: number[];
 	format: FormatType;
 }